Added experimental item code
Largely based on Aura 1, made some minor changes and rewrote picking
stuff up, which was messy before and is a little bit less messy now.
Initially there was a bug with the changing position when broadcasting
entity disappears, item has to be newly generated or removed before it's
position is changed. Or we'd need caching...