Hotfix for StackoverflowException in user data
If you copy a conf file or script list from system to user without
removing the include at the end you get a StackoverflowException because
it's included again and again. This fixes this that problem, but won't
prevent mistakes like including the system file, which would in turn
include the user one again as well.