Fixed quests objectives not being completed
When completing a quest from an NPC we have to complete the objectives,
otherwise the result will be "None", which prevents the player from
getting the rewards registered for "Perfect". Also... it just makes
sense to actually complete the quest if the script says so o.o

This wasn't a problem with quests that are completed manually, with a
click, because those are completely done once you're able to click the
button.

http://aura-project.org/forum/index.php/topic/1140-beginner-quest-school-1-doesnt-give-the-skill-defense