Fixed 64k prop entity id restriction
So far we've increased a counter every time a new server prop was
created, and that counter became the last part of the prop entity id.
This effectively meant you could only create 64k (working) props, which
would've been a problem with dungeons and dynamic regions, where we
constantly create new props, increasing the counter.

Instead, props don't get entity ids on creation anymore now. Instead
it's left as 0, and it's set inside the region when the prop is added.
This way the region controls the ids and we don't have to manage them in
any way.