Added feature G1EasyOverseas
Officially the quests all get duplicated to add a slightly different
reward, like a few more AP in EU. We'll try to do this in the scripts
with a feature, unless it becomes a problem. If it does we should
inherit the quests, so we don't have to copy everything.